Fixes #34350 - Require 'foreman/telemetry' in ldap initializer
